requirements.txt
groq
phidata
yfinance
python-dotenv
duckduckgo-search
sqlalchemy
fastapi[standard]